backend/requirements.txt
fastapi
uvicorn
python-dotenv
scikit-learn
sentence-transformers
requests
textstat
bleach
openai
beautifulsoup4
markdown
slowapi==0.1.9
